+++ EnergyManager/Library/Source/EnergyManager/Utility/Units/Hertz.cpp
@@ -3,7 +3,7 @@
 namespace EnergyManager {
 	namespace Utility {
 		namespace Units {
-			Hertz::Hertz(const unsigned long& value, const SIPrefix& prefix) : SIUnit("Hertz", "Hz", value, prefix) {
+			Hertz::Hertz(const double& value, const SIPrefix& prefix) : SIUnit("Hertz", "Hz", value, prefix) {
 			}
 		}
 	}
