+++ EnergyManager/Library/Source/EnergyManager/Utility/Units/Hertz.hpp
@@ -8,14 +8,14 @@ namespace EnergyManager {
 			/**
 			 * Represents a frequency.
 			 */
-			class Hertz : public SIUnit<Hertz, unsigned long> {
+			class Hertz : public SIUnit<Hertz, double> {
 			public:
 				/**
 				 * Creates a new Hertz.
 				 * @param value The frequency.
 				 * @param prefix The SI prefix.
 				 */
-				Hertz(const unsigned long& value = 0, const SIPrefix& prefix = SIPrefix::NONE);
+				Hertz(const double& value = 0, const SIPrefix& prefix = SIPrefix::NONE);
 			};
 		}
 	}
